+++ src/main/java/spharos/msg/domain/category/repository/CategoryProductRepositoryCustom.java
@@ -5,6 +5,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import spharos.msg.domain.category.dto.CategoryResponse.CategoryProductDto;
 import spharos.msg.domain.category.dto.CategoryResponse.SubCategory;
+import spharos.msg.domain.category.entity.CategoryProduct;
+import spharos.msg.domain.product.entity.Product;
 
 public interface CategoryProductRepositoryCustom {
 
@@ -14,4 +16,5 @@ public interface CategoryProductRepositoryCustom {
 
     Page<CategoryProductDto> findCategoryProductsById(Long categoryId, Pageable pageable);
 
+    CategoryProduct findByProductWithFetchJoin(Product product);
 }
